refactor(domino-royal): switch arena component to js-only module
